Get delegate working again. Clean up by moving common data types to common.ts.
