+++ src/common.ts
@@ -28,6 +28,7 @@ export class StripeConfigCommon {
 
     // See documentation for STPPaymentConfiguration for using these fields.
     // If left 'undefined' the default value will be used.
+    // TODO: Don't use iOS-specific data types
     companyName: string = undefined;
     requiredBillingAddressFields: STPBillingAddressFields = undefined;
     requiredShippingAddressFields: PKAddressField = undefined;
@@ -36,4 +37,52 @@ export class StripeConfigCommon {
     additionalPaymentMethods: STPPaymentMethodType = undefined;
     createCardSources: boolean = undefined;
     stripeAccount: string = undefined;
+}
+
+export interface StripePaymentMethod {
+    image: any; // TODO: UIImage marshals to ???
+    label: string;
+    templateImage: any;
+}
+
+export interface StripeShippingMethod {
+    amount: number;
+    detail: string;
+    label: string;
+    identifier: string;
+}
+
+export interface StripeShippingMethods {
+    /** Is shipping to the address valid? */
+    isValid: boolean;
+    /** If not valid, an error describing the issue with the address */
+    validationError: string;
+    /** The shipping methods available for the address. */
+    shippingMethods: StripeShippingMethod[];
+    /** The pre-selected (default) shipping method for the address. */
+    selectedShippingMethod: StripeShippingMethod;
+}
+
+export interface StripeAddress {
+    name: string;
+    line1: string;
+    line2: string;
+    city: string;
+    state: string;
+    postalCode: string;
+    country: string;
+    phone: string;
+    email: string;
+}
+
+export interface StripePaymentData {
+    isReadyToCharge: boolean;
+    paymentMethod: StripePaymentMethod;
+    shippingInfo: StripeShippingMethod;
+}
+
+export interface StripePaymentListener {
+    onPaymentDataChanged(data: StripePaymentData): void;
+    onError(errorCode: number, message: string): void;
+    provideShippingMethods(address: StripeAddress): StripeShippingMethods;
 }
